bootstrap.sh bug fix

+++ bootstrap.sh
@@ -1,6 +1,6 @@
 #!/bin/bash -x
 
-[ -d ./m4 ] && mkdir ./m4
+[ ! -d ./m4 ] && mkdir ./m4
 
 aclocal -I m4 && autoreconf -f -i -Wall,no-obsolete 
 
@@ -8,6 +8,6 @@ automake --add-missing
 
 # remove some junk
 rm -rf ./autom4te.cache
-rm ./config.h.in~
+[ -f ./config.h.in~ ] && rm ./config.h.in~
